chore(build): update build to match ngM2 build processes (#342)
* fix(build): sync to ngM2 build processes

* lint and style fixes
* tools and package fixes
* npm process improvements
* support aot, build, test, lint, demo-app tasks
* demo css refactors, lint fixes
* stage-deploy:devapp now deploys aot version

* feat(lib): remove uses of rxjs patch operators

Usages of RxJS operators should use the "longhand" syntax, which avoids polluting the Observable.prototype in userland.
Refactors the entire codebase not to use the patch operators from RxJS, because they pollute the user's setup. Instead, opts into importing the operators directly.

> Refs angular/material2#2622.

* chore(demo-app): no longer fallback to index.html file

Gulp Connect currently always falls back to the index.html file if a resource could not be found. This is not a good idea because this way the browser or Angular is never able to report not-found errors.

* chore(build): more changes base on PR feedback

*  artifact scripts should push empty commits
   > When publishing build artifacts (cdk; material) there will be always a change (regardless if something changed in the package or not). This is because the version will be always appended with the current SHA. But when publishing the docs content this is not the case. Sometimes there won't be any change and Git will exit with error code 1 because empty commits are not allowed.
*  remove unused docs, screenshot, and dashboard tools