fix(ngClass): should properly remove classes without fallback (#995)
In some cases where a fallback `ngClass` is not defined, the ClassDirective had no way of knowing to remove its ngClass on deactivation. 

> Unlike other directives, the `removeStyles` action has no effect on ClassDirective.

By specifying an empty default fallback, we ensure that on deactivation - if no other default is specified - the ClassDirective will correctly fallback to an empty string case.


Fixes #992.