build: avoid incorrect provider definition renaming (#1191)
NGC already supports lazy providers and generates definitions for
these providers. Previously those definitions were called
`ngInjectableDef`. With recent changes in Angular, these definition
members have been renamed to `ɵprov`. This conflicts with the legacy
build system that was originally designed for View Engine. The build
system / package-tools rename this member to be unique.

The logic for renaming private members generated by the NGC flat module
bundler exists because individual secondary entry-points are re-exported
in the primary entry-point, and we want to avoid symbol conflicts there.

To fix this in a more reasonable way (same concept used in
`@angular/bazel`), we specify a unique prefix for each entry-point by
leveraging the `flatModulePrivateSymbolPrefix` option.