build: do not include secondary entry-points in primary UMD bundle (#1226)
The APF specification does not capture all corner cases, but conceptually, generated UMD
bundles should match the actual entry-point in the typings tree. This means that it's technically
spec-compliant to include secondary entry-points in a primary UMD bundle. At the same time though,
it's also spec-compliant to not include secondary entry-points and require UMD consumers to load the
secondary entry-point UMD bundles manually. 

We are going with the latter since that helps with an issue that has surfaced in ngcc. The plan is to update the
APF specification to explicitly guard against cases like that, so that primary UMD bundles no longer can bundle
secondary entry-points. This change prepares flex-layout for the change to the APF specification.