fix(show-hide): use initial value as fallback instead of parent
When SSR creates media screen styles, it sets the default `display`
CSS property to `parent.display`. However, if the parent has its
`display` property unset, it results in an empty display that breaks
the correct functionality as the next closest match is made.

This is fixed by setting `display` to `initial` as default instead
of an empty value.

Credit to @vmasek

Fixes #1163