fix(show-hide): set explicit display fallback for SSR
In the event that a user specifies two breakpoints with conflicting
show/hide, e.g. fxHide.lt-lg and fxShow.md, fxShow will not be
able to re-enable CSS `display`, since there's no equivalent to
nulling out the styles, which is what we do when handling it
inline in the browser.

This sets the property to `initial` in SSR stylesheets only. This
should preserve browser behavior where nulling out the styles still
works.

NOTE: if you have other classes that set the `display` property of
an element, they may get overridden in SSR only. This is
unavoidable, since getComputedStyle is not available to use in SSR.

Fixes #1250