fix(core): reset current value when directive is cleared
When a directive is fired, it sets a current value to compare to
future values, that way we don't fire mutiple costly element
change events per media update. However, when a directive moves
to an "uncovered" segment, ie a segment where there is no active
value, we are not clearing the previous current value. This leads
to a lack of updates when styles are not present but should be,
e.g. when a directive moves into an uncovered segment, and then
back into a covered one.

Therefore, we reset the currentValue for a directive when we clear
styles, to enable updates when moving back to a covered segment
later.