fix(core): only emit changed events from MediaObserver
It is unfortunately the case in a real-world browser environment
that overlapping breakpoints trigger separate events in our change
monitoring schema. This is (nearly) unavoidable, but what we can
do is ensure that we only emit distinct events for our event
aggregator, MediaObserver. We now only emit when we get an entirely
distinct set of media changes from the previous state.