fix(server): disable breakpoints correctly and avoid style overuse
This problem is two-fold. First, on the server, when we deactivated
breakpoints, we were deactivating them in the wrong order. This
meant that after a breakpoint was activated, its styles would
linger, polluting the correct state of the style map.

The second problem was overpopulation of styles. Essentially,
we would provide a fallback style for every single breakpoint
if available. This is redundant, because we already populate a
global activation state at the beginning of style population.

To resolve these issues, we ensure that we deactivate breakpoints
correctly, and we also disable fallback styles on the server after
the initial population is complete.