feat(core): add value multiplication suffix feature
It may be desirable for users to multiply the base values to keep
some sort of internal consistency. For instance, in Material
Design, values are typically multiplied by 4x or 8x. To support
this, we add a new configuration option to the library that plugs
in to select directives (at this point, only `fxFlexOffset` and
`fxLayoutGap`). To use this feature, use the following syntax:

```ts
@NgModule({
  imports: [
    FlexLayoutModule.withConfig({
      multiplier: {
        value: 4,
        unit: 'px',
      }
    }),
  ],
})
export class MyAppModule {}
```

```html
<div fxFlexOffset="3x">My content</div>
```

This yields:

```html
<div style="margin-right: 12px">My content</div>
```

This feature is disabled by default, so to use it, you must
explicitly specify both the value and unit of your multiplication.