feat(core): add support for default unit for unitless values
If a user specifies a value for a directive without a unit, either
intentionally or otherwise, we currently error out for most cases.
Therefore, we introduce a config option that by default appends
a unit to unitless values. The built-in default is `px`, but could
just as easily be set to `%` or `vh` depending on a user's app.

Currently, the only supported directive for this feature is
`fxLayoutGap`, but _not_ in `grid` mode.

To use, follow this syntax:

```ts
@NgModule{{
  imports: [
    FlexLayoutModule.withConfig({defaultUnit: '%'}),
  ],
})
export class MyAppModule {}
```

Then use the following HTML:

```html
<div fxLayoutGap="5">
  <div>Line one</div>
  <div>Line two</div>
</div>
```

This would result in the following markup:

```html
<div fxLayoutGap="5">
  <div style="margin-right: 5%">Line one</div>
  <div>Line two</div>
</div>
```