feat(flex): detect display precedence in fxLayout directive
In some cases, a user may intentionally hide a directive using
styling that exists outside of the scope of this library. However,
fxLayout had no way of detecting this, or establishing a precedence
for pre-existing styles. This adds detection and special caching
for pre-styled display attributes on an element. This means that if,
at any point in the change detection of the directive for its entire
lifespan, it detects that the style has been set to `display: none`,
it will use that as the display value in place of what it otherwise
would have used.

Please note: this means that you need to be more careful when using
this functionality! If, for instance, you have functionality that
sets `display: none` on a directive, that functionality needs to
now take over restoring the correct style for the directive. This
is essentially an escape hatch from the library: saying that you
know better how to manage these directives.

To use this feature, do the following:

```ts
@NgModule({
  imports: [
    FlexLayoutModule.withConfig({detectLayoutDisplay: true}),
  ],
})
export class AppModule {}
```

Your markup should then adapt automatically to the new behavior.