fix(private): remove fallback from flex-direction auto-prefixer (#1408)
When we clear styles on the server, we emit an empty value in
order to flush the cache. However, all values still end up
passing through the auto-prefixer, which then converted this
value to the default value of row. However, other components
own the default 'row' value fallback, and so this only serves
to add another layer of logic that gets in the way. Now, we remove
it so that the overlying mechanisms can work as intended in both
browser and server.

Fixes #1394